Suporte a produtos. Listar produtos ainda causa erros.

+++ src/main/java/br/ufsm/inf/bolicho/dao/NewUserDAO.java
@@ -20,7 +20,7 @@ public class NewUserDAO extends NewGenericDAO<User> {
         update(user);
     }
 
-    public void excluir(long id) {
+    public void excluir(int id) {
         User u = findById(id);
         delete(u);
     }
